Revert "PR for welcome screen"
