--- app/screens/Auth/register.js
@@ -1,17 +1,0 @@
-import React, { Component } from 'react';
-import {
-  Text,
-  View,
-} from 'react-native';
-
-
-export default class Register extends Component {
-  render() {
-    return (
-      <View>
-        <Text>This is the Register Page</Text>
-      </View>
-    );
-  }
-}
-
