+++ package.json
@@ -14,7 +14,6 @@
 		"babel-jest": "20.0.3",
 		"babel-preset-react-native": "3.0.2",
 		"jest": "20.0.4",
-		"react-navigation": "^1.0.0-beta.11",
 		"react-test-renderer": "16.0.0-alpha.12"
 	},
 	"jest": {
